Fetch labels only if module is not provided
