+++ hub/integrations/cleanlab/__init__.py
@@ -1,3 +1,2 @@
 from .cleanlab import clean_labels
 from .cleanlab import clean_view
-from .cleanlab import create_tensors
